Fullscreen

+++ khafile.js
@@ -1,10 +1,12 @@
 let project = new Project('My Project');
 
+const fs = require('fs');
+
 const BUILD_DIR = 'build';
 const INDEX_HTML = 'index.html';
 
+project.addShaders('shaders/**');
 project.addAssets('assets/**.png');
-project.addAssets('shaders/**');
 project.addSources('source');
 project.addParameter('-dce full'); // Dead Code Elimination (Haxe, do your thing!)
 project.buildPath = BUILD_DIR;
@@ -12,15 +14,13 @@ project.buildPath = BUILD_DIR;
 if(platform === Platform.HTML5) {
     project.addDefine('js-es6');
 }
-
-resolve(project);
-
 function copy_callback(err) {
     if (err) throw err;
 }
 
-if (platform == Platform.HTML5) {
-    const fs = require('fs');
-    const BUILD_INDEX = BUILD_DIR + '/html5/' + INDEX_HTML;
+const BUILD_INDEX = BUILD_DIR + '/html5/' + INDEX_HTML;
+if (platform == Platform.HTML5 && fs.existsSync(BUILD_INDEX)) {
     fs.copyFile(INDEX_HTML, BUILD_INDEX, copy_callback);
 }
+
+resolve(project);
